Added an online booking category
In this commit I added an online booking system, complete with optional payment systems; This implementaion is obviously a front-end to a back-end that could be developed later in the future. The implementation isn't complete and is built strictly for desktop clients. As well as a start on the work to implement a Naql App section